Work started moving converters to individual files.
